requirements.txt
Flask==2.0.3
numpy==1.19.5
omegaconf==2.3.0
Pillow==8.4.0
requests==2.27.1
torch==1.10.1
torchvision==0.11.2